phinx-187: fix incorrect existing test on dual-column PK.
The order of columns in a key is significant, even though it
does not affect the uniqueness test (which is obviously vital
for primary keys). For example, the PK is no good for
searching based on just the second column in the index - it
can't be used. Therefore index tests must treat two indexes
with the same columns in a different order as being unique.